requirements.txt
# Core ML libraries
numpy>=1.24.0
pandas>=2.0.0
scikit-learn>=1.3.0
imbalanced-learn>=0.12.0
xgboost>=1.7.0
lightgbm>=4.0.0

# Deep learning (optional)
# torch>=2.0.0
# tensorflow>=2.13.0

# Data visualization
matplotlib>=3.7.0
seaborn>=0.12.0
plotly>=5.14.0

# Dimensionality reduction / representation learning
umap-learn>=0.5.5

# Model evaluation and interpretation
shap>=0.42.0
lime>=0.2.0

# Experiment tracking
mlflow>=2.5.0

# Data validation
great-expectations>=0.17.0

# Hyperparameter optimization
optuna>=3.2.0

# Utilities
joblib>=1.3.0
tqdm>=4.65.0
python-dotenv>=1.0.0
pyyaml>=6.0.0

# Testing
pytest>=7.4.0
pytest-cov>=4.1.0

# Code quality
black>=23.7.0
flake8>=6.0.0
mypy>=1.4.0

# Jupyter notebook support
jupyter>=1.0.0
ipywidgets>=8.0.0

# RAG stack
chromadb>=0.5.5
sentence-transformers>=3.0.0
